implement uath-hook in App.js
